requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
cors==1.0.1
Django==3.2.16
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.11.0
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
PySocks==1.7.1
python-decouple==3.8
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
urllib3==1.26.15
zope.event==4.6
zope.interface==6.0
